finance/requirements.txt
Flask
Flask-Session
passlib
mysql-connector-python
